Rename references to 'Terragrunt' to 'state'
Terragrunt is a tool but the logic we implement around it is to
enable us to manage state.
