+++ cdflow_commands/plugins/ecs.py
@@ -21,7 +21,7 @@ from cdflow_commands.logger import logger
 from cdflow_commands.plugins import Plugin
 from cdflow_commands.plugins.base import Destroy as BaseDestroy
 from cdflow_commands.secrets import get_secrets
-from cdflow_commands.terragrunt import (
+from cdflow_commands.state import (
     S3BucketFactory, LockTableFactory, initialise_terraform_backend
 )
